Do not install numpy from a .deb package
When installing it from a .deb package, it pins its version to that package, which can be a fairly old one. Implicitly adding it to python-packages (numpy is a dependency of ifcopenshell) allows us to use a more modern version. In the future, it can also be added explicitly there, with particular version to pin it to if necessary.